Create Unmanned_raliway_crossing
This code is designed to obtain the vibration data from a vibration sensor and raise an alarm if the vibration is detected beyond a certain threshold for a certain amount of time. The vibration sensor is attached to the track of the train and whenever a train approaches, the vibrations in the track are used to detect its presence.