Create Water_level_warning_system
This project is designed to raise an alarm whenever the level of the fluid, say water, fills up beyond a desired threshold. A switch which gets closed whenever there is a tilt in its axis beyond a certain degree is used for the project. The switch is mounted on an arm that can move freely in the vertical direction and this arm is attached to the inner side of the bucket. The output of the switch is used to switch on the actuator.